Move popup code to separate module
