+++ plugins/SBOMVis/customTemplates/template.html
@@ -134,6 +134,10 @@
             var network;
         </script>
 
+        <script type="inline-module" id="popupModule">
+            {% include 'src/popup.js' %}
+        </script>
+
         <script type="inline-module" id="buttonEventHandlersModule">
             {% include 'src/buttonEventHandlers.js' %}
         </script>
